+++ src/main/java/roomescape/global/exception/AuthorizationException.java
@@ -1,4 +1,4 @@
-package roomescape.globalException;
+package roomescape.global.exception;
 
 public class AuthorizationException extends RuntimeException {
