+++ src/main/java/roomescape/global/exception/ErrorResponse.java
@@ -1,0 +1,7 @@
+package roomescape.global.exception;
+
+public record ErrorResponse(
+        String message
+) {
+}
+
