+++ src/main/java/roomescape/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.globalException;
+package roomescape.global.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
