--- src/main/java/roomescape/globalException/ErrorResponse.java
@@ -1,7 +1,0 @@
-package roomescape.globalException;
-
-public record ErrorResponse(
-    String message
-) {
-}
-
