+++ src/main/java/roomescape/theme/exception/InvalidThemeException.java
@@ -1,6 +1,6 @@
 package roomescape.theme.exception;
 
-import roomescape.globalException.InvalidInputException;
+import roomescape.global.exception.InvalidInputException;
 
 public class InvalidThemeException extends InvalidInputException {
