+++ src/main/java/roomescape/user/exception/InvalidUserException.java
@@ -1,6 +1,6 @@
 package roomescape.user.exception;
 
-import roomescape.globalException.InvalidInputException;
+import roomescape.global.exception.InvalidInputException;
 
 public class InvalidUserException extends InvalidInputException {
